+++ Realm/Tests/MigrationTests.mm
@@ -22,7 +22,7 @@ extern "C" {
 #import "RLMMigration.h"
 #import "RLMProperty_Private.h"
 #import "RLMSchema_Private.h"
-#import "RLMObjectStore.h"
+#import "RLMObjectStore.hpp"
 #import "RLMObjectSchema_Private.hpp"
 
 @interface MigrationObject : RLMObject
@@ -67,7 +67,7 @@ extern "C" {
     [RLMRealm migrateRealmAtPath:RLMTestRealmPath() withBlock:^NSUInteger(RLMMigration *migration, NSUInteger oldSchemaVersion) {
         XCTAssertEqual(oldSchemaVersion, 0U, @"Initial schema version should be 0");
         [migration enumerateObjects:MigrationObject.className
-                                       block:^(RLMObject *oldObject, RLMObject *newObject) {
+                              block:^(RLMObject *oldObject, RLMObject *newObject) {
             XCTAssertThrows(oldObject[@"stringCol"], @"stringCol should not exist on old object");
             NSNumber *intObj;
             XCTAssertNoThrow(intObj = oldObject[@"intCol"], @"Should be able to access intCol on oldObject");
@@ -173,7 +173,7 @@ extern "C" {
                                        block:^(RLMObject *oldObject, RLMObject *newObject) {
             NSNumber *intObj = oldObject[@"stringCol"];
             XCTAssert([intObj isKindOfClass:NSNumber.class], @"Old stringCol should be int");
-            newObject[@"stringCol"] = [NSString stringWithFormat:@"%@", intObj];
+            newObject[@"stringCol"] = intObj.stringValue;
         }];
         return 1;
     }];
@@ -181,20 +181,38 @@ extern "C" {
     // verify migration
     realm = [self realmWithTestPath];
     MigrationObject *mig1 = [MigrationObject allObjectsInRealm:realm][1];
-    XCTAssertEqualObjects(mig1[@"stringCol"], @"2", @"strintCol should be string after migration.");
+    XCTAssertEqualObjects(mig1[@"stringCol"], @"2", @"stringCol should be string after migration.");
 }
 
-- (void)testNoMigrationApplied {
-    // create realm an tables
-    RLMRealm *realm = [self dynamicRealmWithTestPathAndSchema:[RLMSchema sharedSchema]];
-    realm = nil;
-    
-    // apply migration
-    XCTAssertNoThrow(
-        [RLMRealm migrateRealmAtPath:RLMTestRealmPath() withBlock:^NSUInteger(__unused RLMMigration *migration, NSUInteger oldSchemaVersion) {
-            return oldSchemaVersion;
-        }],
-        @"Returning the same version should work when no migration is required");
+- (void)testVersionNumberCanStaySameWithNoSchemaChanges {
+    @autoreleasepool { [self dynamicRealmWithTestPathAndSchema:[RLMSchema sharedSchema]]; }
+
+    XCTAssertNoThrow([RLMRealm migrateRealmAtPath:RLMTestRealmPath() withBlock:^NSUInteger(__unused RLMMigration *migration, NSUInteger oldSchemaVersion) {
+        return oldSchemaVersion;
+    }]);
+}
+
+- (void)testVersionNumberMustIncreaseWithSchemaChanges {
+    @autoreleasepool {
+        // make string an int
+        RLMObjectSchema *objectSchema = [RLMObjectSchema schemaForObjectClass:MigrationObject.class];
+        RLMProperty *stringCol = objectSchema.properties[1];
+        stringCol.type = RLMPropertyTypeInt;
+        stringCol.objcType = 'i';
+
+        // create realm with old schema and populate
+        RLMRealm *realm = [self realmWithSingleObject:objectSchema];
+        [realm beginWriteTransaction];
+        RLMCreateObjectInRealmWithValue(realm, MigrationObject.className, @[@1, @1]);
+        [realm commitWriteTransaction];
+    }
+
+    XCTAssertThrows([RLMRealm migrateRealmAtPath:RLMTestRealmPath() withBlock:^NSUInteger(RLMMigration *migration, NSUInteger oldSchemaVersion) {
+        [migration enumerateObjects:MigrationObject.className block:^(RLMObject *, RLMObject *newObject) {
+            newObject[@"stringCol"] = @"";
+        }];
+        return oldSchemaVersion;
+    }]);
 }
 
 @end