+++ CHANGELOG.md
@@ -9,10 +9,15 @@
 
 ### Enhancements
 
+* `[RLMObject initWithObject:]` and `[RLMObject createInRealmWithObject:]` now support
+  any object type with kvc properties.
 * Improve performance when getting the count of items matching a query but not
   reading any of the objects in the results.
 * Add a return value to `-[RLMRealm refresh]` that indicates whether or not
   there was anything to refresh.
+* Add the class name to the error message when an RLMObject is missing a value
+  for a property without a default.
+* Add support for opening Realms in read-only mode.
 
 ### Bugfixes